camera_requirements.txt
Flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
python-dotenv==1.0.0
